backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai     # Gemini SDK
PyMuPDF                 # PDF → text
python-docx             # DOCX → text
Pillow                  # basic image handling
